test(73-02): migrate 7 local_search tests to integration file, delete inline block
- Appended 7 unit tests (improve, factory, config, strategy/mode defaults) to tests/engines/local_search.rs
- Added required imports: factory, factory_with_config, HillClimbingConfig, LinearChromosome, LocalSearchOperator, GaError, RangeChromosome, Cow
- Deleted inline #[cfg(test)] block from src/operations/local_search.rs (lines 271-365)
- All 15 local_search tests pass from integration file
