+++ src/operations/local_search.rs
@@ -268,98 +268,3 @@ impl LocalSearchOperator for HillClimbingConfig {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::chromosomes::Range as RangeChromosome;
-    use crate::genotypes::Range as RangeGene;
-    use crate::traits::{ChromosomeT, LinearChromosome};
-
-    /// Simple quadratic fitness: sum of squares.
-    fn quadratic(dna: &[RangeGene<f64>]) -> f64 {
-        dna.iter().map(|g| g.value * g.value).sum::<f64>()
-    }
-
-    #[test]
-    fn test_hill_climbing_returns_improvements_count() {
-        let genes: Vec<RangeGene<f64>> = (0..5)
-            .map(|i| RangeGene::new(i, vec![(-10.0, 10.0)], 8.0))
-            .collect();
-        let mut chromo = RangeChromosome::<f64>::new();
-        chromo.set_dna(Cow::Owned(genes));
-        chromo.set_fitness(quadratic(chromo.dna()));
-
-        let config = HillClimbingConfig {
-            step_size: 1.0,
-            max_iterations: 50,
-        };
-        let result = config.improve(&mut chromo, &quadratic);
-        assert!(result.is_ok());
-        let improvements = result.unwrap();
-        assert!(improvements > 0, "expected at least one improvement");
-        // Fitness should be lower (minimization) after hill climbing
-        assert!(
-            chromo.fitness() < 400.0,
-            "fitness should improve: {}",
-            chromo.fitness()
-        );
-    }
-
-    #[test]
-    fn test_hill_climbing_unsupported_type() {
-        use crate::chromosomes::Binary as BinaryChromosome;
-        use crate::genotypes::Binary as BinaryGene;
-
-        let mut chromo = BinaryChromosome::new();
-        let result = HillClimbingConfig::default().improve(&mut chromo, &|_: &[BinaryGene]| 0.0);
-        assert!(result.is_err());
-        match result {
-            Err(GaError::LocalSearchError(msg)) => {
-                assert!(msg.contains("HillClimbing"));
-            }
-            _ => panic!("expected LocalSearchError"),
-        }
-    }
-
-    #[test]
-    fn test_hill_climbing_empty_dna() {
-        let mut chromo = RangeChromosome::<f64>::new();
-        let result = HillClimbingConfig::default().improve(&mut chromo, &quadratic);
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), 0);
-    }
-
-    #[test]
-    fn test_local_search_application_strategy_default() {
-        let strategy = LocalSearchApplicationStrategy::default();
-        assert!(matches!(
-            strategy,
-            LocalSearchApplicationStrategy::AllOffspring
-        ));
-    }
-
-    #[test]
-    fn test_local_search_mode_default() {
-        let mode = LocalSearchMode::default();
-        assert!(matches!(mode, LocalSearchMode::Lamarckian));
-    }
-
-    #[test]
-    fn test_factory_returns_enum() {
-        let op = factory(LocalSearch::HillClimbing);
-        assert_eq!(op, LocalSearch::HillClimbing);
-    }
-
-    #[test]
-    fn test_factory_with_config() {
-        let config = HillClimbingConfig {
-            step_size: 0.01,
-            max_iterations: 10,
-        };
-        let result = factory_with_config(LocalSearch::HillClimbing, config);
-        // factory_with_config returns HillClimbingConfig which implements LocalSearchOperator
-        let mut chromo = RangeChromosome::<f64>::new();
-        let improve_result = result.improve(&mut chromo, &|_: &[RangeGene<f64>]| 0.0);
-        assert!(improve_result.is_ok());
-    }
-}
